refactor: move config-meta tests from testRun.ts into separate files
